safari-600.7-branch/LayoutTests/fast/forms/file/file-setrangetext-expected.txt

Test setRangeText() method is not available in file inputs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Running tests on input with attributes: {"type":"file"}

PASS element.value = '0123456789XYZ' threw exception Error: InvalidStateError: DOM Exception 11.
PASS element.setRangeText('ABC', 0, 0) threw exception Error: InvalidStateError: DOM Exception 11.
PASS element.value is ""
PASS successfullyParsed is true

TEST COMPLETE
